+++ app/assets/stylesheets/camp_applications.scss
@@ -5,10 +5,10 @@
 
 .progressbar_lenght
 {
-	width: 19%;
+	width: 19.7%;
 }
 
 .progressbar_seperation_lenght
 {
-	width: 1%;
+	width: .3%;
 }
