--- app/controllers/admins_controller.rb
@@ -1,28 +1,0 @@
-class AdminsController < ApplicationController
-  before_action :set_user, only: %i[edit update show]
-  layout 'admin_layout'
-
-  def show; end
-
-  def edit; end
-
-  def update
-    @admin.skip_password_validation = true
-    if @admin.update(admin_params)
-      bypass_sign_in(@admin)
-      redirect_to admin_path, notice: 'User was successfully updated.'
-    else
-      redirect_to homepage_path, notice: @admin.errors
-    end
-  end
-
-  private
-
-  def set_user
-    @admin = current_admin
-  end
-
-  def admin_params
-    params.require(:admin).permit(:first_name, :middle_name, :last_name, :email, :phone_number, :password, :password_confirmation, :image)
-  end
-end
