+++ app/views/admin_panel/admins/edit.html.erb
@@ -1,0 +1,66 @@
+<div id="main ">
+  <div  class="container mt-5" >
+    <div class="main-content card card-block form-content">
+      <div class="text-center pt-3" >
+      <%=form_with model: [:admin_panel, current_admin], local: true do |f| %>
+       <% if current_admin.errors.any? %>
+          <div id="error_explanation">
+            <h2><%= pluralize(movie.errors.count, "error") %> prohibited this Admin from being saved:</h2>
+            <ul>
+            <% current_admin.errors.full_messages.each do |message| %>
+              <li><%= message %></li>
+            <% end %>
+            </ul>
+          </div>
+        <%end%>      
+        <div class="image-upload" id="profile-pic" >
+          <% if current_admin.image.attached? %>
+              <%= image_tag current_admin.image, :class=> "img-circular"  %> 
+          <% end %>
+          <label for="file-input">
+            <br>
+            <i class="fas fa-camera fa-2x" onclick="uploadImage()"></i>
+          </label>
+          <%= f.file_field :image, id:"file-input" %>
+        </div>
+        <div class="row">
+          <div class="field col-4">
+            <%= f.label :first_name%><br />
+            <%= f.text_field :first_name,autofocus: true, autocomplete: "first_name", class: 'w_100'%>
+          </div>
+          <div class="field col-4">
+            <%= f.label :middle_name%><br />
+            <%= f.text_field :middle_name,autofocus: true, autocomplete: "middle-name", class: 'w_100'%>
+          </div>
+          <div class="field col-4">
+            <%= f.label :last_name%><br />
+            <%= f.text_field :last_name, autofocus: true, autocomplete: "last-name", class: 'w_100'%>
+          </div>
+        </div>
+        <div class="row">
+          <div class="field col-6">
+            <%= f.label :email%><br />
+            <%= f.email_field :email, autofocus: true, autocomplete: "email", class: 'w_100'%>
+          </div>
+          <div class="field col-4" style="margin: 0 auto">
+            <%= f.label :phone_number %><br />
+            <%= f.number_field :phone_number,autofocus: true, autocomplete: "phone-number", class:'quantity w_100'%>
+          </div>
+        </div>
+        <br>
+        <div class="row">
+        <div class="field col-6">
+          <%= f.label :password %>      
+          <%= f.password_field :password, class: 'w_100'%>
+        </div>
+        <div class="field col-6">
+          <%= f.label :password_confirmation %><br />
+          <%= f.password_field :password_confirmation, class: 'w_100'%>
+        </div>
+      </div>           
+        <%= f.submit :Submit ,class:'btn btn-success'%>
+      </div>
+      <%end%>
+    </div>
+  </div>
+</div>
