+++ app/views/admin_panel/admins/show.html.erb
@@ -1,0 +1,1 @@
+<%= render 'admin_panel/home_page_admins/show', user: current_admin %>
