+++ app/views/admin_panel/camp_locations/_camp_list.html.erb
@@ -1,0 +1,23 @@
+<div class="table-responsive pa_20">
+  <table id="mytable" class="table table-bordred table-striped">
+    <thead>
+      <th><%= sortable "name"%></th>
+      <th>Actions</th>
+    </thead>
+    <tbody>
+      <% @camp_locations.each do |camp_location| %>
+        <tr>
+          <td><%= camp_location.name %></td>
+          <td>
+            <%= link_to '<button class="btn buttons"><i class="material-icons table_icons">visibility</i></button>'.html_safe, admin_panel_camp_location_path(camp_location), title: 'Show Camp Location' %>
+            <%= link_to '<button class="btn buttons"><i class="material-icons table_icons">create</i></button>'.html_safe, edit_admin_panel_camp_location_path(camp_location), title: 'Edit Camp Location' %>
+            <%= link_to '<button class="btn buttons"><i class="material-icons table_icons">delete</i></button>'.html_safe,admin_panel_camp_location_path(camp_location), method: :delete, data: { confirm: 'Are you sure?' }, title: 'Delete Camp Location'   %>
+          </td>
+        </tr>
+      <%end%>
+    </tbody>
+  </table>
+  <%= paginate @camp_locations %>
+<%= link_to 'Add New Location' ,new_admin_panel_camp_location_path,class: 'w_color btn btn-success'%>
+<%= link_to 'Export CSV', export_csv_admin_panel_camp_locations_path(format: "csv"),  class: 'w_color btn btn-success' %>
+</div>
