+++ app/views/admin_panel/camp_locations/_form.html.erb
@@ -1,4 +1,4 @@
-<%= form_with model: @camp_location, local: true do |f| %>
+<%= form_with model: [:admin_panel, @camp_location], local: true do |f| %>
   <div class="col-12">
     <%= render "shared/flash_message"%>
   </div>
@@ -8,7 +8,7 @@
   </div>
   <div class="row ml-1 mr-1">
   	<div class="col-6">
-      <%= link_to 'Back', camp_locations_path, class: 'w_100 g_color any_links btn border border-success'%>
+      <%= link_to 'Back', admin_panel_camp_locations_path, class: 'w_100 g_color any_links btn border border-success'%>
   	</div>
     <div class="actions col-6">
       <%= f.submit "Add Location", class: 'w_100 btn btn-success'  %>
