+++ app/views/admin_panel/camp_locations/edit.html.erb
