+++ app/views/admin_panel/camp_locations/index.html.erb
@@ -12,7 +12,7 @@
             <% end %> 
           </div>
         </div>
-        <%= render 'camp_locations/camp_list', camp_locations: @camp_locations %>
+        <%= render 'admin_panel/camp_locations/camp_list', camp_locations: @camp_locations %>
       </div>
     </div>
   </div>
