+++ app/views/admin_panel/camp_locations/new.html.erb
