+++ app/views/admin_panel/camp_locations/show.html.erb
@@ -9,7 +9,7 @@
         <strong>Name:</strong>
         <%= @camp_location.name%>
       </div>
-      <%= link_to 'Back' ,camp_locations_path,class: 'btn btn-success d-flex justify-content-center'%>
+      <%= link_to 'Back' ,admin_panel_camp_locations_path,class: 'btn btn-success d-flex justify-content-center'%>
     </div>
   </div>
 </div>
