+++ app/views/admin_panel/camps/edit.html.erb
