+++ app/views/admin_panel/camps/index.csv.erb
