+++ app/views/admin_panel/camps/new.html.erb
