+++ app/views/admin_panel/camps/show.html.erb
@@ -1,0 +1,56 @@
+<p id="notice"><%= notice %></p>
+
+<p>
+  <strong>Camp Status:</strong>
+  <%= @camp.camp_status %>
+</p>
+<p>
+  <strong>Camp title:</strong>
+  <%= @camp.camp_title %>
+</p>
+
+<p>
+  <strong>Camp type:</strong>
+  <%= @camp.camp_type %>
+</p>
+
+<p>
+  <strong>Applicant registration start date:</strong>
+  <%= @camp.applicant_registration_start_date %>
+</p>
+
+<p>
+  <strong>Applicant registration end date:</strong>
+  <%= @camp.applicant_registration_end_date %>
+</p>
+
+<p>
+  <strong>Applicant registration start time:</strong>
+  <%= @camp.applicant_registration_start_time %>
+</p>
+
+<p>
+  <strong>Applicant registration end time:</strong>
+  <%= @camp.applicant_registration_end_time %>
+</p>
+
+<p>
+  <strong>Parent application start date:</strong>
+  <%= @camp.parent_application_start_date %>
+</p>
+
+<p>
+  <strong>Parent application end date:</strong>
+  <%= @camp.parent_application_end_date %>
+</p>
+
+<p>
+  <strong>Parent application start time:</strong>
+  <%= @camp.parent_application_start_time %>
+</p>
+
+<p>
+  <strong>Parent application end time:</strong>
+  <%= @camp.parent_application_end_time %>
+</p>
+<%= link_to 'Back' ,admin_panel_camps_path, class: 'w_color btn btn-success'%>
