+++ app/views/admin_panel/camps/toggle_status.js.erb
