+++ app/views/admin_panel/home_page_admins/index.html.erb
@@ -12,7 +12,7 @@
             <% end %> 
           </div>
         </div>
-        <%= render(partial: 'home_page_admins/list', locals: { user: @users})%>
+        <%= render(partial: 'list', locals: { user: @users})%>
         </div>
       </div>
     </div>
