+++ app/views/admin_panel/users/edit.html.erb
@@ -1,9 +1,8 @@
-<div id="main"> 
-  <div class="container mt-5"> 
-    <div class="main-content card card-block">
-      <div class="text-center p-5">
-        <%= link_to "Back", homepage_path, class: 'btn btn-success w_color'%>
-      <%= form_with model: @user, url:'create',method: :post, local: true do |f| %>
+<div id="main">
+<div  class="container mt-4">
+  <div class="main-content card card-block form-content">
+    <div class="text-center pt-3" >
+      <%= form_with model: [:admin_panel, @user], local: true do |f| %>
        <% if @user.errors.any? %>
           <div id="error_explanation">
             <h2><%= pluralize(movie.errors.count, "error") %> prohibited this movie from being saved:</h2>
@@ -16,7 +15,7 @@
         <% end %>      
         <div class="image-upload" id="profile-pic" >
           <% if @user.image.attached? %>
-              <%= image_tag user.image, :class=> "img-circular"  %> 
+              <%= image_tag @user.image, :class=> "img-circular"  %> 
           <% end %>
           <label for="file-input">
             <br>
@@ -45,17 +44,29 @@
           </div>
           <div class="field col-6">
             <%= f.label :country %><br />
-            <%= f.country_select :country, [], {class: "btn border border-info"}, autofocus: true, autocomplete: "country", format: :with_full_country_name%>
+            <%= f.country_select :country, [@user.country], {class: "btn border border-info"}, autofocus: true, autocomplete: "country"%>
           </div>
         </div>
-          <div class="field col-4 m-auto" >
+          <div class="field col-4" style="margin: 0 auto">
             <%= f.label :phone_number %><br />
             <%= f.number_field :phone_number,autofocus: true, autocomplete: "phone-number", class:'quantity w_100'%>
           </div>
-          <br>         
-        <%= f.submit :Submit ,class:'btn btn-success'%>
+          <br>
+          <div class="row">
+          <div class="field col-6">
+            <%= f.label :password %>         
+            <%= f.password_field :password, class: 'w_100'%>
+          </div>
+          <div class="field col-6">
+            <%= f.label :password_confirmation %><br />
+            <%= f.password_field :password_confirmation, class: 'w_100'%>
+          </div>
+        </div>                   
+        <%= f.submit :Submit ,class:'btn btn-success w_50'%>
       </div>
       <% end %>
+      <br>
+      <%= link_to "Back", admin_panel_homepage_path, class: 'btn btn-success w_color'%>
     </div>
   </div>
 </div>
