+++ app/views/admin_panel/users/show.html.erb
@@ -1,0 +1,1 @@
+<%= render 'admin_panel/home_page_admins/show', user: @user %>
