--- app/views/admins/show.html.erb
@@ -1,1 +1,0 @@
-<%= render 'home_page_admins/show', user: @admin %>
