+++ app/views/camp_applications/_personal_info_step.html.erb
@@ -1,26 +1,15 @@
 <div>
-  <% if @user.errors.any? %>
+  <% if current_user.errors.any? %>
           <div id="error_explanation">
-            <h2><%= pluralize(@user.errors.count, "error") %> Not saved:</h2>
+            <h2><%= pluralize(current_user.errors.count, "error") %> Not saved:</h2>
             <ul>
-            <% @user.errors.full_messages.each do |message| %>
+            <% current_user.errors.full_messages.each do |message| %>
               <li><%= message %></li>
             <% end %>
             </ul>
           </div>
         <% end %>     
-      <%= form_with model: @user, url: camp_application_path(@camp_application), method: :patch do |user| %>
-        <div class="image-upload" id="profile-pic" >
-          <% if @user.image.attached? %>
-            <%= image_tag @user.image, :class=> "img-circular"  %> 
-          <% end %>
-          <label for="file-input">
-            <br>
-            <i class="fas fa-camera fa-2x " onclick="uploadImage()"></i>
-          </label>
-           <%= user.file_field :image, id:"file-input" %>
-        </div>
-
+      <%= form_with model: current_user, url: camp_application_path(@camp_application), method: :patch do |user| %>
         <div class="row">
           <div class="field col-4">
             <%= user.label :first_name%><br />
