+++ app/views/camps/show.html.erb
@@ -1,56 +1,20 @@
-<p id="notice"><%= notice %></p>
+<div class="container col-5 su" >
+  <div class="main-content card card-block">
+    <div class="ptlr_20">
+      <%= link_to "Back", camps_path, class: 'btn btn-success w_color'%>
+        <p>Welcome and thank you for applying to be a participant at <%= @camp.camp_title %>. CONNECT is an international programme for students ages 14 to 17 from around the world which takes place in partnership.
 
-<p>
-  <strong>Camp Status:</strong>
-  <%= @camp.camp_status %>
-</p>
-<p>
-  <strong>Camp title:</strong>
-  <%= @camp.camp_title %>
-</p>
+        The purpose of this application is for the CONNECT Placement Team to get to know you and learn about your interests, passions, and motivations for registering. Before registering, we encourage you to visit our website and learn more about our offerings and the registration process.
 
-<p>
-  <strong>Camp type:</strong>
-  <%= @camp.camp_type %>
-</p>
+        This registration form is for both CONNECT camps that will take place during July 2022:
+        CONNECT Cycle A: 1:30 AM GMT to 7:30 AM GMT
+        CONNECT Cycle B: 11:30 AM GMT to 5:30 PM GMT
 
-<p>
-  <strong>Applicant registration start date:</strong>
-  <%= @camp.applicant_registration_start_date %>
-</p>
+        If you have any questions during the application process, please feel free to email  at info@projectname.com.
 
-<p>
-  <strong>Applicant registration end date:</strong>
-  <%= @camp.applicant_registration_end_date %>
-</p>
-
-<p>
-  <strong>Applicant registration start time:</strong>
-  <%= @camp.applicant_registration_start_time %>
-</p>
-
-<p>
-  <strong>Applicant registration end time:</strong>
-  <%= @camp.applicant_registration_end_time %>
-</p>
-
-<p>
-  <strong>Parent application start date:</strong>
-  <%= @camp.parent_application_start_date %>
-</p>
-
-<p>
-  <strong>Parent application end date:</strong>
-  <%= @camp.parent_application_end_date %>
-</p>
-
-<p>
-  <strong>Parent application start time:</strong>
-  <%= @camp.parent_application_start_time %>
-</p>
-
-<p>
-  <strong>Parent application end time:</strong>
-  <%= @camp.parent_application_end_time %>
-</p>
-<%= link_to 'Back' ,camps_path, class: 'w_color btn btn-success'%>
+        I AKNOWLEDGE AND AGREE THAT ANY INFORMATION PROVIDED IN THIS APPLICATION WILL BE COLLECTED AND PROCESSED IN ACCORDANCE WITH THE PRIVACY POLICY.
+      </p>
+    </div>
+    <%= link_to "Continue", profile_camp_application_path(@camp_application,@camp), class: 'ml-4 btn btn-success w_color mr-4'%>
+  </div>
+</div>
