--- app/views/camps/show.json.jbuilder
@@ -1,1 +1,0 @@
-json.partial! "camps/camp", camp: @camp
