+++ app/views/devise/sessions/new.html.erb
@@ -1,36 +1,38 @@
-<div class="container col-4 si">
-  <div class="main-content card card-block">
-    <div class="text-center p_20">
-      <h2>Welcome to Camp Manager</h2>
-      <p>Please sign in or create an account</p>
-    </div>
-    <hr/>
-    <div class="card-body login-form">
-      <%= form_for(resource, as: resource_name, url: session_path(resource_name)) do |f| %>
-        <div class="field col-12 ">
-          <%= f.label :email, "Email Address:*"%><br />
-          <%= f.email_field :email, autofocus: true, autocomplete: "email", class: 'w_100' %>
-        </div>
-        <div class="field col-12 ">
-          <%= f.label :password, "Password:*" %><br />
-          <%= f.password_field :password, autocomplete: "current-password", class: 'w_100' %>
-        </div>
-        <% if devise_mapping.rememberable? %>
-          <div class="field ml-3">
-            <%= f.check_box :remember_me%>
-            <%= f.label :remember_me %>
+<div>
+  <div class="d-flex justify-content-end p-4 "> <%= link_to 'Admin Panel' ,admin_panel_homepage_path,class: ' btn-sm btn-dark text-light'%></div>
+  <div class="container col-4 si">
+    <div class="main-content card card-block">
+      <div class="text-center p_20">
+        <h2>Welcome to Camp Manager</h2>
+        <p>Please sign in or create an account</p>
+      </div>
+      <hr/>
+      <div class="card-body login-form">
+        <%= form_for(resource, as: resource_name, url: session_path(resource_name)) do |f| %>
+          <div class="field col-12 ">
+            <%= f.label :email, "Email Address:*"%><br />
+            <%= f.email_field :email, autofocus: true, autocomplete: "email", class: 'w_100' %>
+          </div>
+          <div class="field col-12 ">
+            <%= f.label :password, "Password:*" %><br />
+            <%= f.password_field :password, autocomplete: "current-password", class: 'w_100' %>
+          </div>
+          <% if devise_mapping.rememberable? %>
+            <div class="field ml-3">
+              <%= f.check_box :remember_me%>
+              <%= f.label :remember_me %>
+            </div>
+          <% end %>
+          <div class="actions col-12">
+            <%= f.submit "Sign In", class: 'btn btn-success w_100'  %>
+            <a href="<%= new_user_password_path %>" class="any_links g_color m-110" >Forget your password?</a>
+            <p class="divider-text">
+              <span class="bg-light">OR</span>
+            </p>
+            <a href="<%= new_user_registration_path %>" class="w_100 text-white btn btn-success" role="button" aria-disabled="true">Sign Up</a>
           </div>
         <% end %>
-        <div class="actions col-12">
-          <%= f.submit "Sign In", class: 'btn btn-success w_100'  %>
-          <a href="<%= new_user_password_path %>" class="any_links g_color m-110" >Forget your password?</a>
-          <p class="divider-text">
-            <span class="bg-light">OR</span>
-          </p>
-          <a href="<%= new_user_registration_path %>" class="w_100 text-white btn btn-success" role="button" aria-disabled="true">Sign Up</a>
-        </div>
-      <% end %>
+      </div>
     </div>
   </div>
 </div>
-
