+++ app/views/devise_admin/sessions/new.html.erb
@@ -1,35 +1,38 @@
-<div class="container" style="width: 40rem; margin-top: 10%;">
-  <div class="main-content card card-block">
-    <div class="card-body text-center">
+<div>
+  <div class="d-flex justify-content-end p-4 "> <%= link_to 'User Login' ,camps_path,class: ' btn-sm btn-dark text-light'%></div>
+  <div class="container" style="width: 40rem; margin-top: 10%;">
+    <div class="main-content card card-block">
+      <div class="card-body text-center">
 
-      <%= form_for(resource, as: resource_name, url: session_path(resource_name)) do |f| %>       
-        <%= image_tag("logo" , id:'logo') %>
-       
-        <h1 class="h3 mb-3 font-weight-normal">Please SignIn</h1>
-        <h1 class="h5 mb-3 font-weight-normal">to Admin Panel</h1>
+        <%= form_for(resource, as: resource_name, url: session_path(resource_name)) do |f| %>       
+          <%= image_tag("logo" , id:'logo') %>
+         
+          <h1 class="h3 mb-3 font-weight-normal">Please SignIn</h1>
+          <h1 class="h5 mb-3 font-weight-normal">to Admin Panel</h1>
 
-        <div class="field">
-          <%= f.label :email , class: "sr-only", for: "inputEmail" %><br />
-          <%= f.email_field :email, autofocus: true, autocomplete: "email", class: "form-control", id: "inputEmail" %>
-        </div>
-
-        <div class="field">
-          <%= f.label :password %><br />
-          <%= f.password_field :password, autocomplete: "current-password" , class: "form-control" %>
-        </div>
-
-        <% if devise_mapping.rememberable? %>
           <div class="field">
-            <%= f.check_box :remember_me %>
-            <%= f.label :remember_me %>
+            <%= f.label :email , class: "sr-only", for: "inputEmail" %><br />
+            <%= f.email_field :email, autofocus: true, autocomplete: "email", class: "form-control", id: "inputEmail" %>
+          </div>
+
+          <div class="field">
+            <%= f.label :password %><br />
+            <%= f.password_field :password, autocomplete: "current-password" , class: "form-control" %>
+          </div>
+
+          <% if devise_mapping.rememberable? %>
+            <div class="field">
+              <%= f.check_box :remember_me %>
+              <%= f.label :remember_me %>
+            </div>
+          <%end%>
+
+          <div class="actions">
+            <%= f.submit "Log in", class: "btn btn-lg btn-primary btn-block" %>
           </div>
         <%end%>
-
-        <div class="actions">
-          <%= f.submit "Log in", class: "btn btn-lg btn-primary btn-block" %>
-        </div>
-      <%end%>
-      <%= render "devise_admin/shared/links" %>
+        <%= render "devise_admin/shared/links" %>
+      </div>
     </div>
   </div>
 </div>
