--- app/views/users/index.html.erb
@@ -1,8 +1,0 @@
-hello
-<br><br>
-<%= @user.username %>
-<%= link_to '<button class="btn side_border buttons"><i class="material-icons table_icons">logout</i></i></button>'.html_safe,destroy_user_session_path, :method => 'delete'  %>
-<br>
-<br>
-<br>
-<%= link_to 'Camps', camps_path ,class: 'btn btn-success'%>
