--- app/views/users/profile.html.erb
@@ -1,79 +1,0 @@
-<div class="container col-7 su" >
-  <div class="main-content card card-block">
-    <div class="ptlr_20">
-      <%= link_to "Back to All Camps", camps_path, class: 'btn btn-success w_color'%>
-      <h4 class="text-center">You have completed <%#= @user.percentage%>0% of your application</h4>
-      <p class="fs">Hi <%= @user.username %>!<br><br>
-      	Thank you for starting a registration form for CONNECT!<br><br>
-      	This is your personal dashboard. You can log into this portal to check the status of your registration form, view a copy of your registration form, and check the status of your parent/gaurdian form.<br><br>
-      	If you have any question, feel free to<a href="#" class="any_links g_color"> Conatct Us.</a><br><br>
-      	Thank you again for you submission!
-      </p>
-    	<%= link_to 'Continue to Complete Application', edit_user_path(current_user.id), class: 'w_color btn btn-success'%>
-    	<hr>
-    </div>
-    <div class="row .bg-light" >
-    	<div class="col-9 ml-4">
-    		<p> Personal Information </p>
-    	</div>
-    	<div class="col-2 mb">
-    	  <p class="w_color w_100 btn btn-success"> Done </p>
-    	</div>
-    </div>
-    <div class="row .bg-light" >
-    	<div class="col-9 ml-4">
-    		<p> Contact Information </p>
-    	</div>
-    	<div class="col-2 mb">
-        <p class="w_color w_100 btn btn-success"> Done </p>
-      </div>
-    </div><div class="row .bg-light" >
-    	<div class="col-9 ml-4">
-    		<p> Social Media Information </p>
-    	</div>
-    	<div class="col-2 mb">
-        <p class="w_color w_100 btn btn-success"> Done </p>
-      </div>
-    </div>
-    <div class="row .bg-light" >
-    	<div class="col-9 ml-4">
-    		<p> Virtual Camp Preference </p>
-    	</div>
-    	<div class="col-2 mb">
-        <p class="w_color w_100 btn btn-success"> Done </p>
-      </div>
-    </div>
-    <div class="row .bg-light" >
-    	<div class="col-9 ml-4">
-    		<p> Educational Information </p>
-    	</div>
-    	<div class="col-2 mb">
-        <p class="w_color w_100 btn btn-success"> Done </p>
-      </div>
-    </div>
-    <div class="row .bg-light" >
-    	<div class="col-9 ml-4">
-    		<p> Previous Camps </p>
-    	</div>
-    	<div class="col-2 mb">
-        <p class="w_color w_100 btn btn-success"> Done </p>
-      </div>
-    </div>
-    <div class="row .bg-light" >
-    	<div class="col-9 ml-4">
-    		<p> Technology Requirement </p>
-    	</div>
-    	<div class="col-2 mb">
-        <p class="w_color w_100 btn btn-success"> Done </p>
-      </div>
-    </div>
-    <div class="row .bg-light" >
-    	<div class="col-9 ml-4">
-    		<p> Enrich Activities </p>
-    	</div>
-    	<div class="col-2 mb">
-        <p class="w_color w_100 btn btn-success"> Done </p>
-      </div>
-    </div>
-	</div>
-</div>
