--- app/views/users/update.js.erb
@@ -1,1 +1,0 @@
-$("#personal_info_tag").css('background-color', 'red');
